[Editorial] Proofread - hyphens for words broken across page edge
These words have a hyphen in the page scans because they are broken across one line and the next. So it's ambiguous whether a hyphen was intended by the author, or if it's a printing constraint. I've added or removed hyphens in these situations based on how the words are spelled elsewhere in the book, in non-ambiguous situations.